Refactor move wrapper classes into their own files
