+++ ironflow/gui/workflows/canvas_widgets/base.py
@@ -18,7 +18,7 @@ Number = Union[int, float]
 if TYPE_CHECKING:
     from ironflow.gui.workflows.canvas_widgets.flow import FlowCanvas
     from ironflow.gui.workflows.screen import WorkflowsGUI
-    from ironflow.model.wrappers import Flow
+    from ironflow.model.flow import Flow
 
 
 class CanvasWidget(ABC):
