+++ ironflow/gui/workflows/canvas_widgets/flow.py
@@ -22,7 +22,7 @@ from ironflow.gui.workflows.canvas_widgets.ports import PortWidget
 if TYPE_CHECKING:
     from ironflow.gui.workflows.canvas_widgets.base import Number
     from ironflow.gui.workflows.screen import WorkflowsGUI
-    from ironflow.model.wrappers import Flow
+    from ironflow.model.flow import Flow
     from ironflow.model.node import Node
